Add dynamic labels
Dynamically create labels using the fo.DynamicEmbeddedDocument object.
The label itself is determined by the suffix of the group (e.g. T1w,
Bold, etc.), but this can be changed to reflect the datatype or any
other BIDS entity as desired. Additionally, the sublabels are determined
by the header of the metric file read, thus making it more robust to any
QC metric so long as the contents are tab-seperated (e.g. .tsv) with
metrics as column headers, and individual subject values per row.