Added tests for generating metrics
Since metrics is a separated out module from the pipeline, thought it
would be good to test that the functionality performs as expected. Here,
we also added a bids2table call using the test data to avoid having to
mock data and metrics.

Additionally, added a type ignore for pkg_resources as mypy complains
about missing stub files (untyped). Don't want to globally ignore this
as it is useful.