Refactor participant pipeline to components (#10)
* Fix worker conditional

* Refactor __main__

- Group together analyis level options in the arg parse.
- Move argparse to internal function to make the main function more
  legible
- Move dumping of dataset_description into the internal function,
  passing the output directory as needed
- Switch from if-elif-else statements to py3.10+ match-case statements
- Renamed `mriqc_dir` variable to `qc_dir` to be more "general"

* Move metric functions to its own module

- Renamed functions to better reflect their purpose

* Reorganize pipeline imports

Moved group and participant_raw import into the niftyone.pipelines.__init__.py
to allow for cleaner import and calls directly from niftyone.pipelines.

* Seperate figure generation from participant code

Moved the generation of the figures to its own submodule, separated by
modality.

* Fix launch call in unit tests

- Also removed note in code
  -> Should the figure generation (e.g. for loop going through
images.nested) be called by modality rather than by suffix / contrast? This can be
beneficial for both generating the same figures for the same modality,
as well as establishing distinct ones.

* Added tests for generating metrics

Since metrics is a separated out module from the pipeline, thought it
would be good to test that the functionality performs as expected. Here,
we also added a bids2table call using the test data to avoid having to
mock data and metrics.

Additionally, added a type ignore for pkg_resources as mypy complains
about missing stub files (untyped). Don't want to globally ignore this
as it is useful.

* Add unit test for generating figures for raw anat

* Add unit tests for generating bold images

* Skip tests using b2t on github actions.

- Add b2t marker for indicating tests requiring b2t.
- Also add omit calls requiring b2t (in testing) from the coverage

* Change 'participant_raw' to 'participant'

The participant pipeline / functions can be generalized and not just
limited to raw datasets - this change reflects this.

* Move CLI creation to its own module

- Created custom class for CLI
- Seperate functions for different analysis levels
- Updated tool description
- Add unit tests for CLI creation and parsing

* Move metadata into its own module

- Moved "tags" to metadata module.
- Moved dataset_description.json generation to a funcdtion within the
  metadata module.
- Added unit tests to assert file is created and content is expected
  (ideally this would be dynamic, but this works for now).
- Updated imports as needed.

* Rename pipelines -> analysis_levels

- Rename module to better reflect its use
- Moved `launch` to be part of `analysis_levels` module.
- Updated `__init__.py`'s as necessary.

* Fix import to use collections.abc.Sequence