Fix multi-overlay cmaps
Instead of setting a single cmap and cycling through (which only really
works for binary images), the user is to pass a list of color maps to
use, one for each overlay. If there are more overlays than color maps,
warn the user and use the previously set default of 'turbo'. This cause
an issue if there are 2+ overlays don't have corresponding images - this
will also serve as an additional warning for the user when they check
the images.