Refactor plugin code for general view import
Repurposes the code from plugins.py (now deprecated) to register the
different view factories. To maintain compatability, added a "view_name"
attribute which is the name that gets added to the registry.
Additionally, the import code tries to import from the path if a search
path is provided, otherwise will search the sys.path for existing
modules with the "niftyone_" prefix.